foodapp/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
django-rest-passwordreset==1.5.0
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
fuzzywuzzy==0.18.0
idna==3.10
multidict==6.1.0
pillow==11.0.0
propcache==0.2.1
PyJWT==2.10.0
python-decouple==3.8
python-dotenv==1.0.1
reportlab==4.2.5
requests==2.32.3
sqlparse==0.5.2
twilio==9.4.1
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.2.3
user-agents==2.2.0
yarl==1.18.3
